requirements.txt
annotated-types==0.7.0   
blinker==1.8.2
certifi==2024.8.30       
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.10
instagrapi==2.1.2        
itsdangerous==2.2.0      
Jinja2==3.1.4
MarkupSafe==2.1.5        
mysql-connector==2.2.9   
pillow==10.4.0
pycryptodomex==3.20.0    
pydantic==2.7.1
pydantic_core==2.18.2    
PySocks==1.7.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4